requirements.txt
matplotlib~=3.8.2
numpy~=1.26.3
PyYAML==6.0
simpy~=4.1.1
pydotplus~=2.0.2
gymnasium~=0.29.1
torch~=2.1.2+cu121
torchrl~=0.2.1
tqdm~=4.64.1
tensordict~=0.2.1
